Support new SIM card message functions
Read/Delete and Copy from a SIM for SIM card messages are added newly.
Only the primary user can read the SIM card message list. In addition,
When this app is set as default SMS app, it allows to delete the SIM
card messages and copy them to a device storage.

Test: Manual

Change-Id: I71bcd093e4220aaf02656e81021e7bc028ee29fc
Signed-off-by: Taesu Lee <taesu82.lee@samsung.com>