Support the Replace Short Message
If new message has a Replace Short Message Type code, then any existing
stored message having the same Protocol Identifier code and Originating
address is replaced with the new message. If there is no message to be
replaced, the new message will be handled in the normal way.
(See TS 23.040 9.2.3.9)

Test: Manual

Change-Id: I6cff6b0bad1e205a0f006111d3e462b66d193f64
Signed-off-by: Taesu Lee <taesu82.lee@samsung.com>