Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> > Patch Set 8: -Code-Review
>  > 
>  >  > Patch Set 8:
>  >  >  
>  >  >  > Patch Set 8:
>  >  >  >  
>  >  >  > In every case, the original authors declare the copyright belongs to AOSP and licensed by AOSP under Apache2. It all seems to be first-party licensed. Was there an item of concern I missed?
>  >  >  
>  >  > Really none, initially it was non aosp Apache 2 header in a file I added, later changed.
>  >  > I think it's mainly because if the size of this external contribution, unusual to see :)
>  > 
>  > If you are not the original author, you cannot arbitrarily change the licensing from something else to Apache2. Which file had a different header, what did it previously say, and who was the original author?
> 
> Oh sorry, Joey ( who signed off ) and I are the authors and he agreed on whatever change I made even before uploading to the aosp Gerrit by signing our CLA
> The original file was a normal Apache 2 java file with LineageOS Project as the author ( my "company" if it helps you ) and I replaced it with Aosp Open Source Project.
> Everything is fine :)

So, you are saying it was originally licensed under Apache2 by the original authors. If, that's the case, I don't think there is any need to change the copyright owner declaration in the first place, but if all of the original authors/contributors have signed the CLA and if all of the original authors/contributors agree to the change, I don't see any harm in doing so.

If you want, we can ask opensource-licensing@google.com to make sure.

Patch-set: 8
Reviewer: Gerrit User 1359095 <1359095@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0