Update patch set 8
Patch Set 8:

Hi Max,

The reason I added license-escalation, which added you, relates to the only added file, NotificationsUtil.java

I have not gone to the LineageOS gerrit server to verify, but as I understand it, the file was originally developed there under an Apache 2.0 license claiming copyright owned by LineageOS. In copying the file into AOSP, Luca has changed the header to one claiming copyright owned by AOSP. Luca has suggested that the only contributors to the file are Luca and Joey, who is CC'd as a reviewer on this change.

Luca pushed this change under an individual CLA with Google. I found no CLA for Joey.

Are we able to accept this file with the changed copyright header on the basis of Luca's assertions regarding its provenance and the representations Luca made by executing the individual CLA?

LineageOS appears to be another OS project so it's not clear to me whether there is another incorporated partner organization involved where we might ask for a corporate CLA instead. The purpose and number of people who signed off are not clear to me either; although, I understand vaguely that the Linux folks use something similar in lieu of a CLA.

Please advise.

Patch-set: 8
Reviewer: Gerrit User 1359095 <1359095@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0