Revert "Remove abandoned implementation."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/03915343-8199-4a99-ac7a-bc9664c16131). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 364631215
Change-Id: Iae8a819eccb0b6f402975c7287de357c313d62c3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com