Revert "Remove abandoned implementation."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/04797631-cef4-4f3c-84ab-a3affc15b2f9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 364626927
Change-Id: I5de98bd442eb97655cf70a08bc0b9bf1c4c992ef
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com