Revert "Remove abandoned implementation."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/90140507-33f4-4785-aa49-6ce28bb242cb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 364689690
Change-Id: I92db94456799e7e9bb3d7dea9ce98be7771d48e9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com