Revert "Remove abandoned implementation."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c7cc78d9-3cf1-4cee-b4ad-4a8c3562f1fb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 364707224
Change-Id: Ica468f8b6d004d3902c6db8501d1681246df94f9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com