Revert "Remove abandoned implementation."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/05b51fcf-2f66-4954-bf69-0f33f604c928). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 364722542
Change-Id: I3183641e4facfc56e6672e4c16c28428a9854bb9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com