Revert "Remove abandoned implementation."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d0bec66e-c1f9-47ae-a683-eef353b2567c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 364707224
Change-Id: Ibf81d5f0e283144fdf7b1e66696e1d3e01c53f15
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com